# src/main/webapp/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "EGE Webservice",
    "description": "EGE Webservice API to convert, validate and customize MEI related data.",
    "contact": {
      "name": "Anne Ferger",
      "url": "https://github.com/Edirom/MEIGarage",
      "email": "anne.ferger@upb.de"
    },
    "license": {
      "name": " GPL-3.0 license",
      "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
    },
    "version": "1.0"
  },
...
  "paths": {
...
    "/ege-webservice/Validation": {
      "get": {
        "tags": [
          "ege-webservice"
        ],
        "summary": "Get all available validations",
        "description": "Return list of input data types and lists of possible validation paths",
        "operationId": "doGet_3",
        "responses": {
          "200": {
            "description": "List of possible validations is returned",
            "content": {
              "text/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Validations"
                }
              }
            }
          },
          "405": {
            "description": "Wrong method error message if the method is called wrong"
          }
        }
      }
    },
    "/ege-webservice/Validation/{input-document-type}": {
      "post": {
        "tags": [
          "ege-webservice"
        ],
        "summary": "Do validation",
        "description": "Validate files of different data formats",
        "operationId": "doPost_2",
        "parameters": [
          {
            "name": "input-document-type",
            "in": "path",
            "description": "Input document type",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/DataInputType"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "fileToConvert": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Validation Result",
            "content": {
              "text/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResult"
                }
              }
            }
          },
          "405": {
            "description": "Wrong method error message if the method is called wrong"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Validations": {
        "type": "object",
        "xml": {
          "name": "validations"
        },
        "properties": {
          "InputDataType": {
            "type": "object",
            "xml": {
              "name": "input-data-type"
            },
            "properties": {
              "id": {
                "type": "string",
                "xml": {
                  "attribute": true
                },
                "example": "Documents:TEI P5 XML Document:TEI,text/xml"
              },
              "xlink-href": {
                "type": "string",
                "xml": {
                  "attribute": true
                },
                "example": "https://teigarage.tei-c.de/ege-webservice/Validation/TEI%3Atext%3Axml/"
              }
            }
          }
        }
      },
      "DataInputType": {
        "type": "string",
        "example": "TEI:text:xml"
      },
...
      "ValidationResult": {
        "type": "object",
        "xml": {
          "name": "validation-result"
        },
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ERROR",
              "SUCCESS",
              "FATAL"
            ]
          },
          "messages": {
            "type": "array",
            "xml": {
              "name": "messages",
              "wrapped": true
            },
            "items": {
              "type": "string",
              "xml": {
                "name": "message"
              }
            }
          }
        }
      },
...
    }
  }
}